backend/debug_session.py
import requests
import os
import firebase_admin
from firebase_admin import credentials, auth
from dotenv import load_dotenv

load_dotenv()

# We need a valid Firebase ID token to test this endpoint.
# Alternatively, I can temporarily disable auth on the endpoint or mock it out for debugging.

# Let's temporarily mock out the auth dependency in session.py to just return a dummy user.
# But actually, doing exactly what debug.md says:
# "If `ephemeral_token` is "dummy_ephemeral_token" or "dummy_fallback_token_due_to_error" — the token minting is failing silently. This means the frontend is connecting to Gemini Live with a fake token and Gemini rejects it immediately, causing the WebSocket to close. If this is the case: go to `backend/gemini.py`, find the `mint_ephemeral_token` function, add print statements inside every except block to print the full traceback, restart the backend, run the curl again, and paste what prints."
