backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
firebase-admin>=6.5.0
google-cloud-firestore>=2.15.0
google-cloud-storage>=2.14.0
google-cloud-run>=0.10.5
google-cloud-pubsub>=2.20.0
google-genai>=0.3.0
websockets>=12.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.1
google-cloud-aiplatform>=1.40.0
